+++ database/migrations/2018_09_12_134432_create_emails_table.php
@@ -14,10 +14,10 @@ class CreateEmailsTable extends Migration
     public function up()
     {
         Schema::create('emails', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->string('address');
-            $table->unsignedInteger('type_id');
+            $table->unsignedBigInteger('type_id');
             $table->boolean('is_active')->default(true);
             $table->morphs('emailable');
             $table->timestamps();
