+++ database/migrations/2018_09_12_141130_create_department_jobtitle_table.php
@@ -14,9 +14,9 @@ class CreateDepartmentJobTitleTable extends Migration
     public function up()
     {
         Schema::create('department_jobtitle', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('department_id');
-            $table->unsignedInteger('jobtitle_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('department_id');
+            $table->unsignedBigInteger('jobtitle_id');
             $table->timestamps();
 
             $table->foreign('department_id')
