+++ database/migrations/2018_09_24_143834_create_customers_table.php
@@ -14,13 +14,13 @@ class CreateCustomersTable extends Migration
     public function up()
     {
         Schema::create('customers', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->string('code')->nullable();
             $table->string('name');
             $table->json('attention')->nullable();
             $table->integer('payment_term')->nullable();
-            $table->unsignedInteger('account_code')->nullable();
+            $table->unsignedBigInteger('account_code')->nullable();
             $table->timestamp('banned_at')->nullable();
             $table->timestamps();
             $table->softDeletes();
