+++ database/migrations/2018_09_29_145145_create_department_employee_table.php
@@ -14,8 +14,8 @@ class CreateDepartmentEmployeeTable extends Migration
     public function up()
     {
         Schema::create('department_employee', function (Blueprint $table) {
-            $table->unsignedInteger('department_id');
-            $table->unsignedInteger('employee_id');
+            $table->unsignedBigInteger('department_id');
+            $table->unsignedBigInteger('employee_id');
             $table->timestamp('joined_at')->nullable();
             $table->timestamp('left_at')->nullable();
             $table->timestamps();
