+++ database/migrations/2018_09_29_230856_create_employee_family_table.php
@@ -14,10 +14,10 @@ class CreateEmployeeFamilyTable extends Migration
     public function up()
     {
         Schema::create('employee_family', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('employee_id');
-            $table->unsignedInteger('family_id');
-            $table->unsignedInteger('relationship_type');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('employee_id');
+            $table->unsignedBigInteger('family_id');
+            $table->unsignedBigInteger('relationship_type');
             $table->timestamps();
             $table->softDeletes();
