+++ database/migrations/2018_10_15_150101_create_licenses_table.php
@@ -14,11 +14,11 @@ class CreateLicensesTable extends Migration
     public function up()
     {
         Schema::create('licenses', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->string('code');
             $table->string('name');
-            $table->unsignedInteger('regulator')->nullable();
+            $table->unsignedBigInteger('regulator')->nullable();
             $table->timestamps();
             $table->softDeletes();
