+++ database/migrations/2018_11_07_103328_create_aircraft_taskcard_table.php
@@ -14,9 +14,9 @@ class CreateAircraftTaskcardTable extends Migration
     public function up()
     {
         Schema::create('aircraft_taskcard', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('taskcard_id');
-            $table->unsignedInteger('aircraft_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('taskcard_id');
+            $table->unsignedBigInteger('aircraft_id');
             $table->timestamps();
             $table->softDeletes();
