+++ database/migrations/2018_12_26_152114_create_quotations_table.php
@@ -14,22 +14,22 @@ class CreateQuotationsTable extends Migration
     public function up()
     {
         Schema::create('quotations', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->string('number')->nullable();
             $table->string('title')->nullable();
-            $table->unsignedInteger('project_id');
+            $table->unsignedBigInteger('project_id');
             $table->json('attention')->nullable();
             $table->timestamp('requested_at')->nullable();
             $table->timestamp('valid_until')->nullable();
-            $table->unsignedInteger('currency_id');
+            $table->unsignedBigInteger('currency_id');
             $table->double('exchange_rate');
             $table->double('subtotal')->nullable();
             $table->json('charge')->nullable();
             $table->integer('ppn')->nullable();
             $table->boolean('is_ppn')->nullable();
             $table->double('grandtotal')->nullable();
-            $table->unsignedInteger('scheduled_payment_type')->nullable();
+            $table->unsignedBigInteger('scheduled_payment_type')->nullable();
             $table->json('scheduled_payment_amount')->nullable();
             $table->string('term_of_payment')->nullable();
             $table->text('term_of_condition')->nullable();
