+++ database/migrations/2019_01_09_135559_create_eo_instructions_table.php
@@ -14,16 +14,16 @@ class CreateEOInstructionsTable extends Migration
     public function up()
     {
         Schema::create('eo_instructions', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
-            $table->unsignedInteger('taskcard_id')->nullable();
-            $table->unsignedInteger('work_area')->nullable();
+            $table->unsignedBigInteger('taskcard_id')->nullable();
+            $table->unsignedBigInteger('work_area')->nullable();
             $table->unsignedDecimal('estimation_manhour', 8, 2)->nullable();
             $table->integer('engineer_quantity')->default(1);
             $table->integer('helper_quantity')->nullable();
             $table->boolean('is_rii')->default(false);
             $table->unsignedDecimal('performance_factor', 8, 2)->nullable();
-            $table->unsignedInteger('sequence')->nullable();
+            $table->unsignedBigInteger('sequence')->nullable();
             $table->longText('description')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
