+++ database/migrations/2019_01_10_140502_create_eo_item_table.php
@@ -14,11 +14,11 @@ class CreateEoItemTable extends Migration
     public function up()
     {
         Schema::create('eo_item', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('eo_id');
-            $table->unsignedInteger('item_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('eo_id');
+            $table->unsignedBigInteger('item_id');
             $table->double('quantity');
-            $table->unsignedInteger('unit_id');
+            $table->unsignedBigInteger('unit_id');
             $table->text('note')->nullable();
             $table->timestamps();
             $table->softDeletes();
