+++ database/migrations/2019_02_04_175701_create_purchase_orders_table.php
@@ -14,21 +14,21 @@ class CreatePurchaseOrdersTable extends Migration
     public function up()
     {
         Schema::create('purchase_orders', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->string('number')->nullable();
-            $table->unsignedInteger('vendor_id');
-            $table->unsignedInteger('purchase_request_id');
+            $table->unsignedBigInteger('vendor_id');
+            $table->unsignedBigInteger('purchase_request_id');
             $table->timestamp('ordered_at')->nullable();
             $table->timestamp('valid_until')->nullable();
             $table->text('shipping_address');
             $table->timestamp('ship_at')->nullable();
-            $table->unsignedInteger('currency_id');
+            $table->unsignedBigInteger('currency_id');
             $table->double('exchange_rate')->nullable();
             $table->double('total_before_tax')->nullable();
             $table->double('tax_amount')->nullable();
             $table->double('total_after_tax')->nullable();
-            $table->unsignedInteger('top_type');
+            $table->unsignedBigInteger('top_type');
             $table->integer('top_day_amount')->nullable();
             $table->timestamp('top_start_at')->nullable();
             $table->text('description')->nullable();
