+++ database/migrations/2019_02_14_131540_create_item_quotation_table.php
@@ -14,11 +14,11 @@ class CreateItemQuotationTable extends Migration
     public function up()
     {
         Schema::create('item_quotation', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('item_id');
-            $table->unsignedInteger('quotation_id');
-            $table->unsignedInteger('taskcard_id');
-            $table->unsignedInteger('pricelist_unit_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('item_id');
+            $table->unsignedBigInteger('quotation_id');
+            $table->unsignedBigInteger('taskcard_id');
+            $table->unsignedBigInteger('pricelist_unit_id');
             $table->double('pricelist_price');
             $table->double('subtotal');
             $table->text('note')->nullable();
