+++ database/migrations/2019_05_24_143917_create_project_workpackage_facilities_table.php
@@ -14,11 +14,11 @@ class CreateProjectWorkPackageFacilitiesTable extends Migration
     public function up()
     {
         Schema::create('project_workpackage_facilities', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
-            $table->unsignedInteger('project_workpackage_id');
-            $table->unsignedInteger('facility_id');
-            $table->unsignedInteger('price_id')->nullable();
+            $table->unsignedBigInteger('project_workpackage_id');
+            $table->unsignedBigInteger('facility_id');
+            $table->unsignedBigInteger('price_id')->nullable();
             $table->double('price_amount')->nullable();
             $table->text('note')->nullable();
             $table->timestamps();
