+++ database/migrations/2019_06_04_135037_create_approvals_table.php
@@ -14,10 +14,10 @@ class CreateApprovalsTable extends Migration
     public function up()
     {
         Schema::create('approvals', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
             $table->morphs('approvable');
-            $table->unsignedInteger('approved_by');
+            $table->unsignedBigInteger('approved_by');
             $table->timestamps();
             $table->softDeletes();
