+++ database/migrations/2019_06_11_190708_create_defectcard_propose_correction_table.php
@@ -14,9 +14,9 @@ class CreateDefectcardProposeCorrectionTable extends Migration
     public function up()
     {
         Schema::create('defectcard_propose_correction', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('defectcard_id');
-            $table->unsignedInteger('propose_correction_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('defectcard_id');
+            $table->unsignedBigInteger('propose_correction_id');
             $table->string('propose_correction_text')->nullable();
             $table->timestamps();
             $table->softDeletes();
