+++ database/migrations/2019_06_29_132806_create_taskcard_workpackage_successors_table.php
@@ -14,10 +14,10 @@ class CreateTaskCardWorkPackageSuccessorsTable extends Migration
     public function up()
     {
         Schema::create('taskcard_workpackage_successors', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->char('uuid', 36)->unique();
-            $table->unsignedInteger('taskcard_workpackage_id');
-            $table->unsignedInteger('next');
+            $table->unsignedBigInteger('taskcard_workpackage_id');
+            $table->unsignedBigInteger('next');
             $table->integer('order');
             $table->timestamps();
             $table->softDeletes();
